+++ examples/motomed_example.py
@@ -1,9 +1,9 @@
 import time
-from pyScienceMode2.rehastim_interface import Stimulator2
+from pyScienceMode2 import Rehastim2
 
 port = "/dev/ttyUSB0"  # Enter the port on which the stimulator is connected
 
-motomed = Stimulator2(port, show_log=True, with_motomed=True).motomed
+motomed = Rehastim2(port, show_log=True, with_motomed=True).motomed
 motomed.init_phase_training(arm_training=True)
 print(motomed.get_motomed_mode())
 motomed.start_phase(speed=50, gear=5, active=False, go_forward=False, spasm_detection=True)
