+++ pyScienceMode2/__init__.py
@@ -1,6 +1,8 @@
-from . import motomed_interface
+from .motomed_interface import _Motomed
 from . import sciencemode
-from . import channel
 from . import utils
-from . import rehastim_interface
+from .rehastim2_interface import Rehastim2
+from .rehastimP24_interface import RehastimP24
 from . import acks
+from .channel import Channel
+from .channel import Point
