+++ pyScienceMode2/rehastimP24_interface.py
@@ -1,375 +1,13 @@
-"""
-Stimulator Interface class used to control the rehamove2.
-See ScienceMode2 - Description and protocol for more information.
-"""
-from typing import Tuple
+
 import time
 from pyScienceMode2.acks import *
 from pyScienceMode2 import channel
 from pyScienceMode2.utils import *
 from pyScienceMode2.sciencemode import RehastimGeneric
-from pyScienceMode2.motomed_interface import _Motomed
 from sciencemode_p24 import sciencemode
 
 
-class Stimulator2(RehastimGeneric):
-    """
-    Class used for the communication with Rehastim2.
-    """
-
-    def __init__(self, port: str, show_log: bool = False, with_motomed: bool = False):
-        """
-        Creates an object stimulator.
-
-        Parameters
-        ----------
-        port : str
-            Port of the computer connected to the Rehastim2.
-        show_log: bool
-            If True, the log of the communication will be printed.
-        with_motomed: bool
-            If the motomed is connected to the Rehastim, put this flag to True.
-        """
-        self.list_channels = None
-        self.stimulation_interval = None
-        self.inter_pulse_interval = 2
-        self.low_frequency_factor = 0
-        self.electrode_number = 0
-        self.electrode_number_low_frequency = 0
-
-        self.amplitude = []
-        self.pulse_width = []
-        self.mode = []
-        self.muscle = []
-        self.given_channels = []
-        self.stimulation_started = None
-        self.device_type = "Rehastim2"
-
-        super().__init__(port, show_log, with_motomed, device_type=self.device_type)
-
-        if with_motomed:
-            self.motomed = _Motomed(self)
-
-        # Connect to Rehastim2
-        packet = None
-        while packet is None:
-            packet = self._get_last_ack(init=True)
-
-        self.send_generic_packet("InitAck", packet=self._init_ack(packet[5]))
-
-    def set_stimulation_signal(self, list_channels: list):
-        """
-        Sets or updates the stimulation's parameters.
-
-        Parameters
-        ----------
-        list_channels: list[channel]
-            Contain the channels and their parameters.
-        """
-        self.amplitude = []
-        self.pulse_width = []
-        self.mode = []
-        self.muscle = []
-        self.given_channels = []
-
-        check_list_channel_order(list_channels)
-
-        for i in range(len(list_channels)):
-            self.amplitude.append(list_channels[i].get_amplitude())
-            self.pulse_width.append(list_channels[i].get_pulse_width())
-            self.mode.append(list_channels[i].get_mode())
-            self.given_channels.append(list_channels[i].get_no_channel())
-
-    def _send_packet(self, cmd: str) -> str:
-        """
-        Calls the methode that construct the packet according to the command.
-
-        Parameters
-        ----------
-        cmd: str
-            Command that will be sent.
-
-        Returns
-        -------
-        In the case of an InitAck, return the string 'InitAck'.
-        """
-
-        packet = [-1]
-        if cmd == "GetStimulationMode":
-            packet = packet_construction(self.packet_count, "GetStimulationMode")
-        elif cmd == "InitChannelListMode":
-            packet = self._packet_init_stimulation()
-        elif cmd == "StartChannelListMode":
-            packet = self._packet_start_stimulation()
-        elif cmd == "StopChannelListMode":
-            packet = packet_construction(self.packet_count, "StopChannelListMode")
-        self.motomed_done.set()
-        init_ack = self.send_generic_packet(cmd, packet)
-        if init_ack:
-            return init_ack
-
-    def _calling_ack(self, packet) -> str:
-        """
-        Processing ack from rehastim
-
-        Parameters
-        ----------
-        packet:
-            Packet which needs to be processed.
-
-        Returns
-        -------
-        A string which is the message corresponding to the processing of the packet.
-        """
-        if packet == "InitAck" or packet[6] == 1:
-            return "InitAck"
-        elif packet[6] == self.TypeReha2["GetStimulationModeAck"].value:
-            return get_mode_ack(packet)
-        elif packet[6] == self.TypeReha2["InitChannelListModeAck"].value:
-            return init_stimulation_ack(packet)
-        elif packet[6] == self.TypeReha2["StopChannelListModeAck"].value:
-            return stop_stimulation_ack(packet)
-        elif packet[6] == self.TypeReha2["StartChannelListModeAck"].value:
-            return start_stimulation_ack(packet)
-        elif packet[6] == self.TypeReha2["StimulationError"].value:
-            return rehastim_error(signed_int(packet[7:8]))
-        elif packet[6] == self.TypeReha2["ActualValues"].value:
-            raise RuntimeError("Motomed is connected, so put the flag with_motomed to True.")
-        else:
-            raise RuntimeError(f"Error packet : not understood {packet[6]}")
-
-    def _packet_init_stimulation(self) -> bytes:
-        """
-        Returns the packet for the InitChannelMode.
-        """
-        coded_inter_pulse_interval = int((self.inter_pulse_interval - 1.5) * 2)
-        msb, lsb = self._msb_lsb_main_stim()
-
-        data_stimulation = [
-            self.low_frequency_factor,
-            self.electrode_number,
-            self.electrode_number_low_frequency,
-            coded_inter_pulse_interval,
-            msb,
-            lsb,
-            0,
-        ]
-
-        packet = packet_construction(self.packet_count, "InitChannelListMode", data_stimulation)
-        return packet
-
-    def _packet_start_stimulation(self) -> bytes:
-        """
-        Returns the packet for the StartChannelListMode.
-        """
-        data_stimulation = []
-        for i in range(len(self.amplitude)):
-            msb, lsb = self._msb_lsb_pulse_stim(self.pulse_width[i])
-            data_stimulation.append(self.mode[i])
-            data_stimulation.append(msb)
-            data_stimulation.append(lsb)
-            data_stimulation.append(int(self.amplitude[i]))
-        packet = packet_construction(self.packet_count, "StartChannelListMode", data_stimulation)
-        return packet
-
-    def _msb_lsb_main_stim(self) -> Tuple[int, int]:
-        """
-        Returns the most significant bit (msb) and least significant bit (lsb) corresponding to the main stimulation
-        interval.
-        Main stimulation interval = [0, 2048] ∙ 0.5 ms + 1 ms
-        Main stimulation interval = [1, 1025] ms
-        Note that in the current software version the minimum main stimulation interval is 8 ms.
-        (ScienceMode2 - Description and Protocol, 4.3 Stimulation Commands, InitChannelListMode)
-
-        Returns
-        -------
-        (msb, lsb): tuple
-            MSB and LSB of main stimulation interval
-        """
-        lsb = msb = -1
-        stimulation_interval_coded = (self.stimulation_interval - 1) / 0.5
-        if stimulation_interval_coded <= 255:
-            lsb = stimulation_interval_coded
-            msb = 0
-        elif 256 <= stimulation_interval_coded <= 511:
-            lsb = stimulation_interval_coded - 256
-            msb = 1
-        elif 512 <= stimulation_interval_coded <= 767:
-            lsb = stimulation_interval_coded - 512
-            msb = 2
-        elif 768 <= stimulation_interval_coded <= 1023:
-            lsb = stimulation_interval_coded - 768
-            msb = 3
-        elif 1024 <= stimulation_interval_coded <= 1279:
-            lsb = stimulation_interval_coded - 1024
-            msb = 4
-        elif 1280 <= stimulation_interval_coded <= 1535:
-            lsb = stimulation_interval_coded - 1280
-            msb = 5
-        elif 1536 <= stimulation_interval_coded <= 1791:
-            lsb = stimulation_interval_coded - 1536
-            msb = 6
-        elif 1792 <= stimulation_interval_coded <= 2047:
-            lsb = stimulation_interval_coded - 1792
-            msb = 7
-        elif stimulation_interval_coded == 2048:
-            lsb = 0
-            msb = 8
-
-        return msb, int(lsb)
-
-    @staticmethod
-    def _msb_lsb_pulse_stim(pulse_width: int) -> Tuple[int, int]:
-        """
-        Returns MSB and LSB corresponding to the pulse width given.
-        Range: [0, 500] μs (in current version [20, 500] μs, if (pw < 20) then pw = 20)
-        (ScienceMode2 - Description and Protocol, 4.3 Stimulation Commands, StartChannelListMode)
-
-        Parameters
-        ----------
-        pulse_width: int
-            Pulse width of a signal.
-
-        Returns
-        -------
-        (msb, lsb): tuple
-            MSB and LSB of pulse_width.
-        """
-        msb = lsb = -1
-        if pulse_width <= 255:
-            lsb = pulse_width
-            msb = 0
-        elif 256 <= pulse_width <= 500:
-            lsb = pulse_width - 256
-            msb = 1
-        return msb, lsb
-
-    def init_channel(
-        self,
-        stimulation_interval: int,
-        list_channels: list,
-        inter_pulse_interval: int = 2,
-        low_frequency_factor: int = 0,
-    ):
-        """
-        Initialize the requested channel.
-        Can update stimulation interval if one is given.
-        Can update list_channels if one is iven.
-
-        stimulation_interval: int
-            Period of the main stimulation. [8,1025] ms.
-        list_channels: list[Channel]
-            List containing the channels and their parameters.
-        """
-        if self.stimulation_active:
-            self.end_stimulation()
-
-        check_stimulation_interval(stimulation_interval)
-        check_unique_channel(list_channels)
-        self.stimulation_interval = stimulation_interval
-        self.list_channels = list_channels
-
-        self.inter_pulse_interval = inter_pulse_interval
-        check_inter_pulse_interval(inter_pulse_interval)
-
-        self.low_frequency_factor = low_frequency_factor
-        check_low_frequency_factor(low_frequency_factor)
-
-        # Find electrode_number (according to Science_Mode2_Description_Protocol_20121212 p17)
-        self.electrode_number = calc_electrode_number(self.list_channels)
-        self.electrode_number_low_frequency = calc_electrode_number(self.list_channels, enable_low_frequency=True)
-
-        self.set_stimulation_signal(self.list_channels)
-        self._send_packet("InitChannelListMode")
-        self._get_last_ack()
-
-    def start_stimulation(self, stimulation_duration: float = None, upd_list_channels: list = None):
-        """
-        Update a stimulation.
-        Warning: only the channel that has been initiated can be updated.
-
-        Parameters
-        ----------
-        stimulation_duration: float
-            Time of the stimulation after the update.
-        upd_list_channels: list[channel]
-            List of the channels that will be updated
-        """
-
-        if upd_list_channels is not None:
-            new_electrode_number = calc_electrode_number(upd_list_channels)
-
-            # Verify if the updated channels have been initialised
-            if new_electrode_number != self.electrode_number:
-                raise RuntimeError("Error update: all channels have not been initialised")
-            self.list_channels = upd_list_channels
-            self.set_stimulation_signal(self.list_channels)
-        self._send_packet("StartChannelListMode")
-        time_start_stim = time.time()
-
-        self._get_last_ack()
-        self.stimulation_active = True
-
-        if stimulation_duration is not None:
-            if stimulation_duration < time.time() - time_start_stim:
-                raise RuntimeError("Asked stimulation duration too short")
-            time.sleep(stimulation_duration - (time.time() - time_start_stim))
-            self.pause_stimulation()
-
-    def pause_stimulation(self):
-        """
-        Update a stimulation.
-        Warning: only the channel that has been initiated can be updated.
-        """
-        tmp_amp = self.amplitude
-        self.amplitude = [0] * len(self.list_channels)
-        self._send_packet("StartChannelListMode")
-        self._get_last_ack()
-        self.amplitude = tmp_amp
-
-    def end_stimulation(self):
-        """
-        Stop a stimulation, after calling this method, init_channel must be used if stimulation need to be restarted.
-        """
-        self._send_packet("StopChannelListMode")
-        self._get_last_ack()
-        self.packet_count = 0
-
-    def get_motomed_angle(self) -> float:
-        """
-        Get the angle of the Motomed.
-
-        Returns
-        -------
-        angle: float
-            Angle of the Motomed.
-        """
-        return self.get_angle()
-
-    def get_motomed_speed(self) -> float:
-        """
-        Get the angle of the Motomed.
-
-        Returns
-        -------
-        angle: float
-            Angle of the Motomed.
-        """
-        return self.get_speed()
-
-    def get_motomed_torque(self) -> float:
-        """
-        Get the angle of the Motomed.
-
-        Returns
-        -------
-        angle: float
-            Angle of the Motomed.
-        """
-        return self.get_torque()
-
-
-class StimulatorP24(RehastimGeneric):
+class RehastimP24(RehastimGeneric):
     """
     Class used for the communication with RehastimP24.
     """
@@ -437,7 +75,7 @@ class StimulatorP24(RehastimGeneric):
         packet_number = self.get_next_packet_number()
         sciencemode.lib.smpt_send_get_extended_version(self.device, packet_number)
         if self.show_log is True:
-            print("Command sent to rehastim:", self.TypeRehap24(sciencemode.lib.Smpt_Cmd_Get_Extended_Version).name)
+            print("Command sent to rehastim:", self.RehastimP24Commands(sciencemode.lib.Smpt_Cmd_Get_Extended_Version).name)
         self._get_last_ack()
         ret = sciencemode.lib.smpt_get_get_extended_version_ack(self.device, extended_version_ack)
         fw_hash = f"fw_hash :{extended_version_ack.fw_hash}"
@@ -458,7 +96,7 @@ class StimulatorP24(RehastimGeneric):
         sciencemode.lib.smpt_send_get_device_id(self.device, packet_number)
 
         if self.show_log is True:
-            print("Command sent to rehastim:", self.TypeRehap24(sciencemode.lib.Smpt_Cmd_Get_Device_Id).name)
+            print("Command sent to rehastim:", self.RehastimP24Commands(sciencemode.lib.Smpt_Cmd_Get_Device_Id).name)
 
         self._get_last_ack()
         ret = sciencemode.lib.smpt_get_get_device_id_ack(self.device, device_id_ack)
@@ -482,7 +120,7 @@ class StimulatorP24(RehastimGeneric):
         sciencemode.lib.smpt_send_get_stim_status(self.device, packet_number)
 
         if self.show_log is True:
-            print("Command sent to rehastim:", self.TypeRehap24(sciencemode.lib.Smpt_Cmd_Get_Stim_Status).name)
+            print("Command sent to rehastim:", self.RehastimP24Commands(sciencemode.lib.Smpt_Cmd_Get_Stim_Status).name)
 
         self._get_last_ack()
         ret = sciencemode.lib.smpt_get_get_stim_status_ack(self.device, stim_status_ack)
@@ -507,7 +145,7 @@ class StimulatorP24(RehastimGeneric):
         sciencemode.lib.smpt_send_get_battery_status(self.device, packet_number)
 
         if self.show_log is True:
-            print("Command sent to rehastim:", self.TypeRehap24(sciencemode.lib.Smpt_Cmd_Get_Battery_Status).name)
+            print("Command sent to rehastim:", self.RehastimP24Commands(sciencemode.lib.Smpt_Cmd_Get_Battery_Status).name)
 
         self._get_last_ack()
         ret = sciencemode.lib.smpt_get_get_battery_status_ack(self.device, battery_status_ack)
@@ -529,7 +167,7 @@ class StimulatorP24(RehastimGeneric):
         sciencemode.lib.smpt_send_get_main_status(self.device, packet_number)
 
         if self.show_log is True:
-            print("Command sent to rehastim:", self.TypeRehap24(sciencemode.lib.Smpt_Cmd_Get_Main_Status).name)
+            print("Command sent to rehastim:", self.RehastimP24Commands(sciencemode.lib.Smpt_Cmd_Get_Main_Status).name)
 
         self._get_last_ack()
         ret = sciencemode.lib.smpt_get_get_main_status_ack(self.device, main_status_ack)
@@ -544,7 +182,7 @@ class StimulatorP24(RehastimGeneric):
         ret = sciencemode.lib.smpt_send_reset(self.device, packet_number)
 
         if self.show_log is True:
-            print("Command sent to rehastim:", self.TypeRehap24(sciencemode.lib.Smpt_Cmd_Reset).name)
+            print("Command sent to rehastim:", self.RehastimP24Commands(sciencemode.lib.Smpt_Cmd_Reset).name)
         self._get_last_ack()
 
     def get_all(self):
@@ -616,7 +254,7 @@ class StimulatorP24(RehastimGeneric):
             raise RuntimeError("Low level initialization failed.")
         self.log(
             "Low level initialized",
-            "Command sent to rehastim: {}".format(self.TypeRehap24(sciencemode.lib.Smpt_Cmd_Ll_Init).name),
+            "Command sent to rehastim: {}".format(self.RehastimP24Commands(sciencemode.lib.Smpt_Cmd_Ll_Init).name),
         )
 
         self.get_next_packet_number()
@@ -690,7 +328,7 @@ class StimulatorP24(RehastimGeneric):
             ll_config.packet_number = self.get_next_packet_number()
             sciencemode.lib.smpt_send_ll_channel_config(self.device, ll_config)
             if self.show_log is True:
-                print("Command sent to rehastim:", self.TypeRehap24(sciencemode.lib.Smpt_Cmd_Ll_Channel_Config).name)
+                print("Command sent to rehastim:", self.RehastimP24Commands(sciencemode.lib.Smpt_Cmd_Ll_Channel_Config).name)
             time.sleep(pulse_interval / 1000)
             self._get_last_ack()
             self.check_ll_channel_config_ack()
@@ -737,7 +375,7 @@ class StimulatorP24(RehastimGeneric):
             raise RuntimeError("Low level stop failed.")
         self.log(
             "Low level stopped",
-            "Command sent to rehastim: {}".format(self.TypeRehap24(sciencemode.lib.Smpt_Cmd_Ll_Stop).name),
+            "Command sent to rehastim: {}".format(self.RehastimP24Commands(sciencemode.lib.Smpt_Cmd_Ll_Stop).name),
         )
         self._get_last_ack()
 
@@ -766,7 +404,7 @@ class StimulatorP24(RehastimGeneric):
             raise RuntimeError("Failed to start stimulation")
         self.log(
             "Stimulation initialized",
-            "Command sent to rehastim: {}".format(self.TypeRehap24(sciencemode.lib.Smpt_Cmd_Ml_Init).name),
+            "Command sent to rehastim: {}".format(self.RehastimP24Commands(sciencemode.lib.Smpt_Cmd_Ml_Init).name),
         )
         self._get_last_ack()
 
@@ -820,7 +458,7 @@ class StimulatorP24(RehastimGeneric):
             raise RuntimeError("Failed to start stimulation")
         self.log(
             "Stimulation started",
-            "Command sent to rehastim: {}".format(self.TypeRehap24(sciencemode.lib.Smpt_Cmd_Ml_Update).name),
+            "Command sent to rehastim: {}".format(self.RehastimP24Commands(sciencemode.lib.Smpt_Cmd_Ml_Update).name),
         )
         self._get_last_ack()
 
@@ -863,7 +501,7 @@ class StimulatorP24(RehastimGeneric):
             raise RuntimeError("Failure to stop stimulation.")
         self.log(
             "Stimulation stopped",
-            "Command sent to rehastim: {}".format(self.TypeRehap24(sciencemode.lib.Smpt_Cmd_Ml_Stop).name),
+            "Command sent to rehastim: {}".format(self.RehastimP24Commands(sciencemode.lib.Smpt_Cmd_Ml_Stop).name),
         )
         self._get_last_ack()
         self.stimulation_started = False
